+++ source/CMakeLists.txt
@@ -355,6 +355,7 @@ target_sources(EndlessSkyLib PRIVATE
 	comparators/ByGivenOrder.h
 	comparators/ByName.h
 	comparators/BySeriesAndIndex.h
+	comparators/UUIDComparator.h
 	orders/OrderSet.cpp
 	orders/OrderSet.h
 	orders/OrderSingle.cpp
